requirements.txt
beautifulsoup4==4.14.3
selenium==4.41.0
python-dotenv==1.2.2
